requirements.txt
streamlit>=1.8.0
pandas>=1.3.0
numpy>=1.21.0
boto3>=1.26.0
python-dotenv>=0.19.0
openpyxl>=3.0.9
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-mock>=3.6.1
moto>=2.3.2
ruff>=0.1.5
black>=21.5b2
modelx>=0.18.0
msal>=1.20.0
requests>=2.28.0
